refactor: fix paths in some files
