+++ project-management-app/src/pages/Welcome/index.tsx
@@ -1,6 +1,6 @@
 import Navbar from '../../components/Navbar/Navbar';
 import { Link } from 'react-router-dom';
-import { routerPaths } from '../../constants/common-constants';
+import { routerPaths } from '../../constants/common';
 import { useSelector } from 'react-redux';
 import { TStore } from '../../store/index';
 import { StyledAbout, StyledWrapper, StyledDescription, StyledWelcomeButton } from './styles';
